Stream 26042020 Comments parsing

+++ cmd/models.go
@@ -2,8 +2,8 @@ package main
 
 // Annotation struct
 type Annotation struct {
-	ID      int64   `gorm:"primary_key"`
-	Seconds float64 `json:"seconds" gorm:"not null"`
-	Text    string  `json:"text"`
-	VideoID string  `json:"videoId" gorm:"not null;index"`
+	ID      int64  `gorm:"primary_key"`
+	Seconds int    `json:"seconds" gorm:"not null"`
+	Text    string `json:"text"`
+	VideoID string `json:"videoId" gorm:"not null;index"`
 }
